+++ src/const.js
@@ -1,0 +1,67 @@
+export const TITLES = [
+  'Made For Each Other',
+  'Popeye Meets Sinbad',
+  'Sagebrush Trail',
+  'The Dance Of Life',
+];
+
+export const POSTERS = [
+  'made-for-each-other.png',
+  'popeye-meets-sinbad.png',
+  'sagebrush-trail.jpg',
+  'the-dance-of-life.jpg',
+];
+
+export const DESCRIPTION_SENTENCES = [
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+  'Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra.',
+  'Aliquam id orci ut lectus varius viverra.',
+  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.',
+  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.',
+  'Sed sed nisi sed augue convallis suscipit in sed felis.',
+  'Aliquam erat volutpat.',
+  'Nunc fermentum tortor ac porta dapibus.',
+  'In rutrum ac purus sit amet tempus.',
+];
+
+export const GENRES = [
+  'Comedy',
+  'Horror',
+  'Action',
+  'Drama',
+  'Documental',
+];
+
+export const WRITERS = [
+  'Takeshi Kitano',
+  'J.J. Abrams',
+  'Mister X',
+  'Stephen King',
+];
+
+export const ACTORS = [
+  'Morgan Freeman',
+  'Danila Kozlovsky',
+  'Daniel Redcliff',
+  'John Travolta',
+];
+
+export const AGE_RATINGS = [
+  0,
+  6,
+  16,
+  18,
+];
+
+export const RELEASE_COUNTRIES = [
+  'Finland',
+  'Russian Federation',
+  'USA',
+  'Canada',
+  'India',
+];
+
+export const DIRECTOR = 'Christofer Nolan';
+export const DATE = '2019-05-11T00:00:00.000Z';
+export const RUNTIME = 77;
+export const ALTERNATIVE_TITLE = 'Laziness Who Sold Themselves';
